+++ src/main/java/com/example/SchoolLunchReport/statistics/domain/desired/entity/FoodNameCountDto.java
@@ -1,0 +1,9 @@
+package com.example.SchoolLunchReport.statistics.domain.desired.entity;
+
+public record FoodNameCountDto(
+    String foodName,
+    Long count
+
+) {
+
+}
