+++ src/main/java/com/example/SchoolLunchReport/statistics/support/FeedBackReader.java
@@ -1,6 +1,10 @@
 package com.example.SchoolLunchReport.statistics.support;
 
-import com.example.SchoolLunchReport.statistics.domain.entity.FeedBack;
+import static com.example.SchoolLunchReport.statistics.domain.type.PeriodType.MONTHLY;
+import static com.example.SchoolLunchReport.statistics.domain.type.PeriodType.WEEKLY;
+
+import com.example.SchoolLunchReport.statistics.domain.feedback.entity.FeedBack;
+import com.example.SchoolLunchReport.statistics.domain.type.PeriodType;
 import com.example.SchoolLunchReport.statistics.repository.FeedBackJpaRepo;
 import java.time.LocalDate;
 import java.util.List;
@@ -23,4 +27,19 @@ public class FeedBackReader {
         return feedBackJpaRepo.findByCreatedAt(startDate);
     }
 
+    public List<FeedBack> getFeedBackInBoundary(LocalDate date, PeriodType periodType) {
+
+        LocalDate startDate;
+        LocalDate endDate;
+
+        if (periodType.equals(WEEKLY)) {
+            startDate = WEEKLY.getStartOfPreviousPeriod(date);
+            endDate = WEEKLY.getStartOfThisPeriod(date);
+        } else {
+            startDate = MONTHLY.getStartOfPreviousPeriod(date);
+            endDate = MONTHLY.getStartOfThisPeriod(date);
+        }
+
+        return getFeedBackInBoundary(startDate, endDate);
+    }
 }
